Initial Commit
Initial commit for HIIT to git. Updates from previous unhosted version include a fix for a bug that would cause the timer to fail to stop when instructed to if the rest period was set to zero. Strings have been moved to strings.xml. Web address updated, the help window has been made more colorful and a new icon has been added.